// frontend-angular/src/frontend-client/model/participantsDishEntry.ts
/**
 * Restaurant Import API
 * To use the API please use the Import API Key specified on \"Restaurants\" page
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SideDish } from './sideDish';

export interface ParticipantsDishEntry { 
    id: string;
    dishId: string;
    dishName: string;
    sideDishes: Array<SideDish>;
    price: number;
    comments: string;
}